+++ src/ArrayStore.php
@@ -82,7 +82,7 @@ class ArrayStore implements KeyValueStore
     /**
      * {@inheritdoc}
      */
-    public function has($key)
+    public function exists($key)
     {
         KeyUtil::validate($key);