+++ src/JsonFileStore.php
@@ -110,7 +110,7 @@ class JsonFileStore implements KeyValueStore
     /**
      * {@inheritdoc}
      */
-    public function has($key)
+    public function exists($key)
     {
         KeyUtil::validate($key);