+++ src/PhpRedisStore.php
@@ -111,7 +111,7 @@ class PhpRedisStore implements KeyValueStore
     /**
      * {@inheritdoc}
      */
-    public function has($key)
+    public function exists($key)
     {
         KeyUtil::validate($key);