+++ src/PredisStore.php
@@ -105,7 +105,7 @@ class PredisStore implements KeyValueStore
     /**
      * {@inheritdoc}
      */
-    public function has($key)
+    public function exists($key)
     {
         KeyUtil::validate($key);